Moved handle_delete.
