+++ include/http/handle_delete.hpp
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   handle_delete.hpp                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/05/21 15:06:57 by irychkov          #+#    #+#             */
+/*   Updated: 2025/05/21 15:17:54 by irychkov         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#pragma once
+#include "core/Location.hpp"
+#include "core/Server.hpp"
+#include "http/HttpRequest.hpp"
+#include "http/HttpResponse.hpp"
+#include "http/HttpResponseBuilder.hpp"
+#include "utils/buildFilePath.hpp"
+#include <sys/stat.h>
+#include <unistd.h>
+
+HttpResponse handleDelete(const HttpRequest&, const Server&, const Location&);
